requirements.txt
fastapi
fastapi[standard]
pydantic
starlette
httpx
uvicorn 
psycopg2-binary
sqlalchemy[asyncio] 
alembic 
python-dotenv
asyncpg
redis
python-jose[cryptography]
passlib[bcrypt]